+++ packages/heroku-ci/commands/ci/config-get.js
@@ -3,6 +3,7 @@ const co = require('co')
 const shellescape = require('shell-escape')
 const api = require('../../lib/heroku-api')
 const Utils = require('../../lib/utils')
+const PipelineCompletion = require('../../lib/completions')
 
 function* run (context, heroku) {
   const pipeline = yield Utils.getPipeline(context, heroku)
@@ -40,7 +41,8 @@ module.exports = {
       name: 'pipeline',
       char: 'p',
       hasValue: true,
-      description: 'pipeline'
+      description: 'pipeline',
+      completion: PipelineCompletion
     }
   ],
   run: cli.command(co.wrap(run))