+++ packages/heroku-ci/commands/ci/last.js
@@ -3,6 +3,7 @@ const co = require('co')
 const api = require('../../lib/heroku-api')
 const TestRun = require('../../lib/test-run')
 const Utils = require('../../lib/utils')
+const PipelineCompletion = require('../../lib/completions')
 
 function* run (context, heroku) {
   const pipeline = yield Utils.getPipeline(context, heroku)
@@ -26,7 +27,8 @@ module.exports = {
       name: 'pipeline',
       char: 'p',
       hasValue: true,
-      description: 'pipeline'
+      description: 'pipeline',
+      completion: PipelineCompletion
     }
   ],
   help: 'looks for the most recent run and returns the output of that run',