+++ packages/heroku-ci/commands/ci/open.js
@@ -1,6 +1,7 @@
 const cli = require('heroku-cli-util')
 const co = require('co')
 const Utils = require('../../lib/utils')
+const PipelineCompletion = require('../../lib/completions')
 
 function* run (context, heroku) {
   const pipeline = yield Utils.getPipeline(context, heroku)
@@ -18,7 +19,8 @@ module.exports = {
       name: 'pipeline',
       char: 'p',
       hasValue: true,
-      description: 'pipeline'
+      description: 'pipeline',
+      completion: PipelineCompletion
     }
   ],
   help: 'opens a browser to view the Dashboard version of Heroku CI',