+++ packages/heroku-ci/lib/completions.js
@@ -1,0 +1,3 @@
+const {PipelineCompletion} = require('cli-engine-heroku/lib/completions')
+
+module.exports = PipelineCompletion