+++ packages/heroku-ci/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Add pipeline flag options completion
 
 ## [1.9.5] - 2017-08-08
 ### Added